Paypal payment in checkout (#262)
* :hourglass: checkout with wizard

* :hourglass: cart inprogress

* templated payment method in cart

* payment method inprogress

* insert order email in buyer

* resolving

* :hourglass: order payment through stripe completed :note: need refinement

* integrated template of checkout upto payment method

* review order design integrated

* review order design integrated

* order review half completed

* :sparkles: completed cart with some validations

* :sparkles: flash messages from constants, order_amount from hashes

* stripe error handling

* :sparkles: No products in cart

* tweak

* :sparkles:

+++ app/controllers/buyers/checkout_controller.rb
@@ -1,15 +1,20 @@
-class Buyers::CheckoutController < ApplicationController
+class Buyers::CheckoutController < Buyers::BaseController
+  skip_before_action :authenticate_buyer!
+  # before_action :find_or_create_buyer, only: [:new, :create_checkout]
 
-  def index
-  end
+  def index; end
 
   def new
     @cart = get_cart
-    @sub_total = sub_total
-    @total = total
+    if !@cart.present?
+      flash[:notice] = I18n.t('flash_messages.no_products_are_added_to_card')
+      redirect_to store_front_path
+      return
+    end
+    @order_amount = order_amount
     @order_id = session_order_id
     if @order_id.present?
-      @order = Order.find(@order_id) rescue nil
+      @order = Order.find(@order_id) || nil
       session.delete(:_current_user_order_id) if !@order.present?
       @order = Order.new if !@order.present?
     else
@@ -26,37 +31,121 @@ class Buyers::CheckoutController < ApplicationController
     if @order_id.present?
       @order = Order.find(@order_id).destroy
     end
-    @order = Order.create(order_line_item_params)
+    @buyer = find_or_create_buyer(order_params[:order_detail_attributes][:email])
+    @order = @buyer.orders.create(order_line_item_params)
     session[:_current_user_order_id] = @order.id
+    redirect_to payment_method_path
+  end
+
+  def payment_method
+    @cart = get_cart
+    @order_amount = order_amount
+    if !@cart.present?
+      flash[:notice] = I18n.t('flash_messages.no_products_are_added_to_card')
+      redirect_to store_front_path
+    end
+  end
+
+  def create_payment_method
+    @order_id = session_order_id
+    if @order_id.present?
+      @order = Order.find(@order_id)
+      @buyer = @order.buyer if @order.present?
+      if @order.present? && @buyer.present?
+        Stripe::RetrieveCardFromToken.call(
+          { stripe_token_id: params[:stripeToken], buyer: @buyer }
+        )
+        customer = Stripe::CustomerCreator.call(
+          { stripe_token_id: params[:stripeToken], buyer: @buyer }
+        )
+        if customer.status
+          flash[:notice] = 'Card added successfully.'
+          redirect_to review_order_path
+        else
+          flash[:notice] = customer.errors
+          render :payment_method
+        end
+      end
+    end
+  end
+
+  def review_order
+    @order_id = session_order_id
+    if @order_id.present?
+      @order = Order.find(@order_id)
+      @cart = get_cart
+      @order_amount = order_amount
+      product_ids = @cart.map { |item| item[:product_id] }
+      @products = Product.find(product_ids) || []
+    else
+      flash[:notice] = I18n.t('flash_messages.no_products_are_added_to_card')
+      redirect_to store_front_path
+    end
+  end
+
+  def create_payment
+    @order_id = session_order_id
+    if @order_id.present?
+      @order = Order.find(@order_id)
+      @buyer = @order.buyer if @order.present?
+      if @order.present? && @buyer.present? && @order.order_type != 'paid_order'
+        @order_amount = order_amount
+        payment_method = @buyer.buyer_payment_methods.last
+        payment = Stripe::ChargeCreator.call(
+          { stripe_token_id: payment_method.stripe_token, buyer: @buyer, order: @order, amount: @order_amount[:total] }
+        )
+        if payment.status
+          session.delete(:_current_user_cart)
+          session.delete(:_current_user_order_id)
+          session.delete(:_selected_payment_method)
+          flash[:notice] = I18n.t('flash_messages.order_placed_successfully')
+          redirect_to store_front_path
+        else
+          flash[:notice] = payment.errors
+          render :review_order
+        end
+      end
+    end
   end
 
   private
+
   def get_cart
     session[:_current_user_cart] ||= []
   end
 
+  def order_amount
+    { total: total, sub_total: sub_total }
+  end
+
   def sub_total
     cart = get_cart
-    @sub_total = 0;
+    @sub_total = 0
     cart.each do |item|
-      product = Product.find(item[:product_id].to_i) rescue nil
-        if product.present?
-          @sub_total = @sub_total + (item[:quantity].to_i * product.price.to_f)
-        end
+      product = Product.find(item[:product_id].to_i) || nil
+      @sub_total += (item[:quantity].to_i * product.price.to_f) if product.present?
     end
-    return @sub_total
+    @sub_total.to_f
   end
 
   def total
     cart = get_cart
-    @total = 0;
+    @total = 0
     cart.each do |item|
-      product = Product.find(item[:product_id].to_i) rescue nil
-        if product.present?
-          @total = @total + item[:quantity].to_i * product.price.to_f
-        end
+      product = Product.find(item[:product_id].to_i) || nil
+      @total += item[:quantity].to_i * product.price.to_f if product.present?
     end
-    return @total
+    @total.to_f
+  end
+
+  def find_or_create_buyer(email)
+    @buyer = Buyer.find_by(email: email)
+    if @buyer.nil?
+      @buyer = Buyer.new({ email: order_params[:order_detail_attributes][:email] })
+      @buyer.skip_password_validation = true
+      @buyer.save
+    end
+    @buyer
   end
 
   def session_order_id
